requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-docx==0.8.11
python-multipart==0.0.6
supabase==2.0.0
python-dotenv==1.0.0
docx2pdf==0.1.8
openai==1.3.0
